+++ scripts/network_functions.sh
@@ -1,26 +1,78 @@
-create_reverse_proxy(){
-  if [ "$SET_REVERSE_PROXY" = "true" ]; then
+set_nginx_cfg(){
+  if [ "$SET_NGINX_CFG" = "true" ]; then
     #check for dependencies
     dep=(nginx)
     dependency_check
     #execute operations
     status_msg "Creating Nginx configuration for $1 ..."
-    cat ${HOME}/kiauh/resources/$1_nginx.cfg > ${HOME}/kiauh/resources/$1
-    sudo mv ${HOME}/kiauh/resources/$1 /etc/nginx/sites-available/$1
-    #ONLY FOR MAINSAIL: replace username if not "pi"
-      if [ "$1" = "mainsail" ]; then
-        sudo sed -i "/root/s/pi/${USER}/" /etc/nginx/sites-available/mainsail
-      fi
+    #copy content from resources to the respective nginx config file
+    cat ${SRCDIR}/kiauh/resources/$1_nginx.cfg > ${SRCDIR}/kiauh/resources/$1
+    ##edit the nginx config file before moving it
+    if [ "$SET_LISTEN_PORT" != "$DEFAULT_PORT" ]; then
+      status_msg "Configuring port for $1 ..."
+      #set listen port ipv4
+      sed -i "s/listen\s[0-9]*;/listen $SET_LISTEN_PORT;/" ${SRCDIR}/kiauh/resources/$1
+      #set listen port ipv6
+      sed -i "s/listen\s\[\:*\]\:[0-9]*;/listen \[::\]\:$SET_LISTEN_PORT;/" ${SRCDIR}/kiauh/resources/$1
+    fi
+    #set correct user
+    if [ "$1" = "mainsail" ] || [ "$1" = "fluidd" ]; then
+      sudo sed -i "/root/s/pi/${USER}/" ${SRCDIR}/kiauh/resources/$1
+    fi
+    #moving the config file into correct directory
+    sudo mv ${SRCDIR}/kiauh/resources/$1 /etc/nginx/sites-available/$1
     ok_msg "Nginx configuration for $1 was set!"
-    #remove default config
-    if [ -e /etc/nginx/sites-enabled/default ]; then
-      sudo rm /etc/nginx/sites-enabled/default
+    if [ "$SET_LISTEN_PORT" != "" ]; then
+      ok_msg "$1 listening on port $SET_LISTEN_PORT!"
+    else
+      ok_msg "$1 listening on def-port $DEFAULT_PORT!"
     fi
-    #create symlink for own configs
-    if [ ! -e /etc/nginx/sites-enabled/$1 ]; then
-      sudo ln -s /etc/nginx/sites-available/$1 /etc/nginx/sites-enabled/
-    fi
+    #remove nginx default config
+    [ -e /etc/nginx/sites-enabled/default ] && sudo rm /etc/nginx/sites-enabled/default
+    #create symlink for own sites
+    [ ! -e /etc/nginx/sites-enabled/$1 ] && sudo ln -s /etc/nginx/sites-available/$1 /etc/nginx/sites-enabled/
     restart_nginx
+  fi
+}
+
+read_listen_port(){
+  LISTEN_PORT=$(grep listen /etc/nginx/sites-enabled/$1 | head -1 | sed 's/^\s*//' | cut -d" " -f2 | cut -d";" -f1)
+}
+
+detect_enabled_sites(){
+  #check if there is another UI config already installed
+  #and reads the port they are listening on
+  if [ -e /etc/nginx/sites-enabled/mainsail ]; then
+    SITE_ENABLED="true" && MAINSAIL_ENABLED="true"
+    read_listen_port "mainsail"
+    MAINSAIL_PORT=$LISTEN_PORT
+    #echo "debug: Mainsail listens on port: $MAINSAIL_PORT"
+  else
+    MAINSAIL_ENABLED="false"
+  fi
+  if [ -e /etc/nginx/sites-enabled/fluidd ]; then
+    SITE_ENABLED="true" && FLUIDD_ENABLED="true"
+    read_listen_port "fluidd"
+    FLUIDD_PORT=$LISTEN_PORT
+    #echo "debug: Fluidd listens on port: $FLUIDD_PORT"
+  else
+    FLUIDD_ENABLED="false"
+  fi
+  if [ -e /etc/nginx/sites-enabled/dwc2 ]; then
+    SITE_ENABLED="true" && DWC2_ENABLED="true"
+    read_listen_port "dwc2"
+    DWC2_PORT=$LISTEN_PORT
+    #echo "debug: DWC2 listens on port: $DWC2_PORT"
+  else
+    DWC2_ENABLED="false"
+  fi
+  if [ -e /etc/nginx/sites-enabled/octoprint ]; then
+    SITE_ENABLED="true" && OCTOPRINT_ENABLED="true"
+    read_listen_port "octoprint"
+    OCTOPRINT_PORT=$LISTEN_PORT
+    #echo "debug: OctoPrint listens on port: $OCTOPRINT_PORT"
+  else
+    OCTOPRINT_ENABLED="false"
   fi
 }
